refactor: move auth typeto types folder, change login input to email
