+++ validation/signin.validation.ts
@@ -1,7 +1,7 @@
 import * as yup from 'yup';
 
 export const signInValidationSchema = yup.object().shape({
-  login: yup.string().email('Invalid email address').required('Login is required'),
+  email: yup.string().email('Invalid email address').required('Email is required'),
   password: yup
     .string()
     .required('Password is required')
