+++ validation/signup.validation.ts
@@ -2,7 +2,7 @@ import * as yup from 'yup';
 
 export const signUpValidationSchema = yup.object().shape({
   name: yup.string().required('Name is required').min(3, 'Name must be at least 3 characters long'),
-  login: yup.string().email('Invalid email address').required('Login is required'),
+  email: yup.string().email('Invalid email address').required('Email is required'),
   password: yup
     .string()
     .required('Password is required')
